refactor(channels/discord): P3 — extract rest.rs (outbound REST + reaction URLs)
Move the channel-message REST builders (send_json, send_with_files,
extract_message_id, edit, delete) and the reaction-URL helpers
(encode_emoji_for_discord, discord_reaction_url) into discord/rest.rs. The
interaction callbacks (defer/reject/followup) stay for P4 (interaction.rs).
Drops now-unused Form/Part/Write imports from mod.rs (Write re-added test-local).
Behaviour-neutral: 161 discord tests pass.
